+++ resources/views/details.blade.php
@@ -23,7 +23,7 @@
                         </div>
                         <div class="py-4 sm:py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
                             <dt class="text-sm font-medium text-gray-500">Size</dt>
-                            <dd class="mt-1 text-sm sm:mt-0 sm:col-span-2">{{ Rych\ByteSize\ByteSize::formatBinary($torrent->size) }}</dd>
+                            <dd class="mt-1 text-sm sm:mt-0 sm:col-span-2">{{ Number::fileSize($torrent->size, precision: 2) }}</dd>
                         </div>
                         <div class="py-4 sm:py-5 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6">
                             <dt class="text-sm font-medium text-gray-500">Downloads</dt>